Component does not support at run-time reconfiguration yet so this is kind of a workaround to make sure it does not breaks for reconfiguration requests.
git-svn-id: file:///home/svn/repos/tigase-pubsub/trunk@288 ea752f3b-7626-48e1-ae42-c030457b2254

+++ src/main/java/tigase/pubsub/PubSubComponent.java
@@ -620,6 +620,13 @@ public class PubSubComponent extends AbstractMessageReceiver implements XMPPServ
 	public void setProperties(Map<String, Object> props) {
 		super.setProperties(props);
 
+		if (props.size() == 1) {
+			// If props.size() == 1, it means this is a single property update 
+			// and this component does not support single property change for the rest
+			// of it's settings
+			return;
+		}
+
 		// Release old resources....
 		if (pubsubRepository != null) {
 			pubsubRepository.destroy();
